Working Member Registration
Members can now register successfully into the database. Additionally a general login and member login page have been added as well as basic code that will be used in logging in members. All of this as a model for how the Intern register/login process will be. Database has been patched to require a username and password to be inserted.