Working Intern Registration
Interns are now registered to the database. In addition, any password inputs are now protected so that the user can't see the charecters put in to prevent others from seeing their password. A few other styling changes were made and now everything is linked as intended for the time being.