Login and Terms Changes
Login files were edited a bit. Still aren't functional yet, but they do not have any syntax errors or other errors present when executed now.

Terms styling is now correct with buttons back to the register forms and the home page.

2 Pages for failed login have been added. Once login in working, one of these pages will appear if a user entered incorrect username and password. On that page, it gives the user the change to go back to the specific login page, or to register.

The button on the bottom of home that misaligned is now more centered as intended.