Edits to Logged in Page and Redirecting
Now, if someone is already logged in, they are brought to the logged in page rather than the log in form since it doesnt make sense to log in if you are already logged in. Still works if a member for instance tried to go to the intern login while still logged in. Log out is also working now.