Working Search Engine
The top option on the search page is a search engine that will be able to find a company or several ones based on words put in. If the word or words exists somewhere in the company's name, it will be displayed similiar to the other 2 search methods implemented. If the full company name is input, a individual page showcasing that one company will be shown. If no results exist, the user will be notified.