JobPost and JobView
You can Now Post Jobs as a Member using a register form. Companies now have their own pages on the website with relevant information and in the future, a table of available jobs determined by the number of interns still needed to reach the max number of interns to work on a job.