dropbox api repo + information
Dropbox will be the source where word docs will be dropped off (such as resume or form results). The information for using api / innovation dropbox are included in comments on the interninsert.php page for now. In the future, students may choose to take previous image insertion from the imgbb site and consolidate it into the dropbox folder instead.