Changes To Intern Form / Resume Insert
Students signing up are notified that file storage for resumes will be done with dropbox, and are given a link to sign up prior to the 0auth prompt they will be met with upon going to the next page. 0auth is essential to create a app usage token for users other than the account that made this ap (Innovation Center's Dropbox account). There are still issues regarding the return-url, but the dropbox prompt is working as expected, bringing users to the 0auth2 dropbox box link on their site to sign in and answer the questions on the prompt to get accesss to the Dropbox App FSUInnovation.