Edits to Intern Form / Intern Insert Url
I sent some inquiries on the dropbox site to get some advice on the issue at hand. I tried doing the app auth on chrome and the buttons would not proceed to the next page on the dropbox 0auth api. For chrome, the redirect uri does not work thus breaking intern registration. I also asked a inquiry about the http header issue and am awaiting a response.