filecheck working
The system knows now when to report a error to user regarding no file upload. Still, there is a issue with the file transfer itself. Files are stored in a temp directory. I am looking for a way to maintain the current system structure and be able to store the file after 0auth2 as the post and temp information is losted after the next page is reached following the one you went to through the form tag.