0auth code flow fixes
the code side of the flow is working ok now. Only seeable issue now is the curl for the token using the code. After that, good chance the file upload will be working assuming the syntax for the file upload curl is functional. Also added error codes to success.php regarding issues that come to play in the oauth/resume process in a switch statement.