resume upload working
Token Upload is working for interns. Now the file is uploaded to dropbox using the token successfully. All that is required is for the sharedlink to be returned from the api as well as cases for a file existing with the same name at the directory path to avoid a api error.