File Upload Works Perfectly
Files Uploaded are not only sent to the correct directory of the app which the website verifies (checks for directory and adds it if not created yet) but also uploads word documents and pdf files without any data corruption. Added to edit profile to get ready for event that resume was not uploaded yet. Member oauth token process was implemented successfully as well.