document download edits
Changed it to a interactive webpage since windows doesnt allow users to chmod permissions as far as writing files to the inner directories is concerned. Instead, file will be uploaded to the apache server first, then user manually presses a button on the webpage downloading the file to the user folder themselves. File is now written correctly without any corruption.